@@ -4,7 +4,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>3.3.5</version>
+    <version>3.4.3</version>
     <relativePath />
   </parent>
   <groupId>org.folio</groupId>
@@ -22,11 +22,11 @@
 
   <properties>
     <!-- runtime dependencies -->
-    <folio-service-tools.version>4.1.0</folio-service-tools.version>
-    <folio-spring.version>8.2.0</folio-spring.version>
-    <spring-cloud-starter-openfeign.version>4.1.3</spring-cloud-starter-openfeign.version>
+    <folio-service-tools.version>5.0.0</folio-service-tools.version>
+    <folio-spring.version>9.0.0</folio-spring.version>
+    <spring-cloud-starter-openfeign.version>4.2.0</spring-cloud-starter-openfeign.version>
     <commons-collections4.version>4.4</commons-collections4.version>
-    <mapstruct.version>1.6.2</mapstruct.version>
+    <mapstruct.version>1.6.3</mapstruct.version>
     <marc4j.version>2.9.6</marc4j.version>
     <coffee-boots.version>4.0.0</coffee-boots.version>
 
@@ -48,7 +48,7 @@
     <maven-enforcer-plugin.version>3.5.0</maven-enforcer-plugin.version>
 
     <!-- other properties -->
-    <java.version>17</java.version>
+    <java.version>21</java.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     <openapi.settings.file>${project.basedir}/src/main/resources/swagger.api/settings.yaml</openapi.settings.file>
