@@ -31,11 +31,11 @@ import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.SqlMergeMode;
 
@@ -75,10 +75,10 @@ class LibraryMappingControllerTest extends BaseControllerTest {
   @Autowired
   private LibraryMappingMapper mapper;
 
-  @MockBean
+  @MockitoBean
   private CentralServerService centralServerService;
 
-  @MockBean
+  @MockitoBean
   private InnReachLocationExternalService innReachLocationExternalService;
 
   @Test
