@@ -33,11 +33,11 @@ import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.SqlMergeMode;
 
@@ -78,10 +78,10 @@ class LocationMappingControllerTest extends BaseControllerTest {
   @Autowired
   private LocationMappingMapper mapper;
 
-  @MockBean
+  @MockitoBean
   private CentralServerService centralServerService;
 
-  @MockBean
+  @MockitoBean
   private InnReachLocationExternalService innReachLocationExternalService;
 
 
