@@ -3,12 +3,12 @@ package org.folio.innreach.repository;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
@@ -41,9 +41,9 @@ abstract class BaseRepositoryTest {
     propertyRegistry.add("spring.datasource.username", postgresqlContainer::getUsername);
   }
 
-  @MockBean
+  @MockitoBean
   private FolioExecutionContext folioExecutionContext;
-  @MockBean
+  @MockitoBean
   private UsersClient usersClient;
 
   @TestConfiguration
